tools/Serial_Checker/requirements.txt
netmiko>=4.2.0
openpyxl>=3.1.2
pandas>=2.0.0
python-dotenv>=1.0.0 